pyproject.toml
[project]
name = "flipkart-product-recommendation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.4.1",
    "flask>=3.1.2",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "langchain>=1.0.5",
    "langchain-astradb>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.5",
    "langchain-groq>=1.0.1",
    "langchain-huggingface>=1.0.1",
    "langchain-ollama>=1.0.0",
    "notebook>=7.4.7",
    "pandas>=2.3.3",
    "prometheus-client>=0.23.1",
    "pypdf>=6.2.0",
    "python-dotenv>=1.2.1",
]
